* feature/go_to_last_msg_merge:
  fix(center.ch.menu): fix ch menu confirm   - private topic에서 leave시 confirm 창 띄우도록 수정
  feature(center.ch.menu): Add ch menu confirm   - #JND-1043 #resolve   - private topic에서 leave시 confirm 창 띄우도록 수정
  feature(center.delete.topic.modal): Add delete topic modal   - template 들여쓰기 수정
  feature(center.delete.topic.modal): Add delete topic modal   - controller, directive 구현

+++ client/app/center/center.controller.js
@@ -60,42 +60,55 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
     loadingTimer : false // no longer using.
   };
 
+  $scope.updateLeftPanel = updateLeftPanel;
+
   $scope.isOwner = function() {
     return ($rootScope.currentEntity.ch_creatorId || $rootScope.currentEntity.pg_creatorId) == memberService.getMemberId();
   };
   $scope.isDefaultTopic = function() {
     return _isDefaultTopic();
   };
-  $scope.onLeaveClick = function() {
-    log('-- leaving')
+  $scope.onLeaveClick = function(channelType) {
+    var isLeaveChannel;
 
-    entityheaderAPIservice.leaveEntity($scope.currentEntity.type, $scope.currentEntity.id)
-      .success(function(response) {
-        log('-- good')
-        // analytics
-        var entity_type = analyticsService.getEntityType($scope.currentEntity.type);
+    log('-- leaving');
 
-        analyticsService.mixpanelTrack( "Entity Leave" , { "type": entity_type } );
-        updateLeftPanel();
-      })
-      .error(function(error) {
-        alert(error.msg);
-      })
+    isLeaveChannel = channelType === 'privategroups' ? confirm($filter('translate')('@ch-menu-leave-private-confirm')) : true;
+
+    if (isLeaveChannel) {
+      entityheaderAPIservice.leaveEntity($scope.currentEntity.type, $scope.currentEntity.id)
+        .success(function(response) {
+          log('-- good');
+          // analytics
+          var entity_type = analyticsService.getEntityType($scope.currentEntity.type);
+
+          analyticsService.mixpanelTrack( "Entity Leave" , { "type": entity_type } );
+          updateLeftPanel();
+        })
+        .error(function(error) {
+          alert(error.msg);
+        });
+    }
   };
-  $scope.onDeleteClick = function() {
-    entityheaderAPIservice.deleteEntity($scope.currentEntity.type, $scope.currentEntity.id)
-      .success(function() {
-        // analytics
-        var entity_type = analyticsService.getEntityType($scope.currentEntity.type);
-        analyticsService.mixpanelTrack( "Entity Delete", { "type": entity_type } );
 
-        updateLeftPanel();
-        fileAPIservice.broadcastChangeShared();
-      })
-      .error(function(error) {
-        alert(error.msg);
-      });
+  $scope.onDeleteTopicClick = function () {
+    if (confirm($filter('translate')('@ch-menu-delete-confirm'))) {
+      entityheaderAPIservice.deleteEntity($scope.currentEntity.type, $scope.currentEntity.id)
+        .success(function() {
+          $scope.updateLeftPanel();
+
+          fileAPIservice.broadcastChangeShared();
+
+          // analytics
+          var entity_type = analyticsService.getEntityType($scope.currentEntity.type);
+          analyticsService.mixpanelTrack("Entity Delete", { "type": entity_type });
+        })
+        .error(function(error) {
+          console.log(error.msg);
+        });
+    }
   };
+
   $scope.onMeesageLeaveClick = function(entityId) {
     $rootScope.$broadcast('leaveCurrentChat', entityId);
   };
@@ -845,22 +858,18 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
         controller  : 'fileUploadModalCtrl',
         size        : 'lg'
       });
-    }
-    else if (selector == 'rename') {
+    } else if (selector == 'rename') {
       $modal.open({
         scope       :   $scope,
         templateUrl :   'app/modal/rename.html',
         controller  :   'renameModalCtrl',
         size        :   'lg'
       });
-    }
-    else if (selector == 'invite') {
+    } else if (selector == 'invite') {
       publicService.openInviteToCurrentEntityModal($scope);
-    }
-    else if (selector == 'inviteUserToChannel') {
+    } else if (selector == 'inviteUserToChannel') {
       publicService.openInviteToJoinedEntityModal($scope);
-    }
-    else if (select == 'share') {
+    } else if (selector == 'share') {
 
     }
   };
